backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
neo4j==5.24.0
openai==1.43.0
python-dotenv==1.0.1
pydantic==2.8.2
httpx==0.27.2